refactor chatconversation
