+++ src/components/Conversation/ChatConversation.vue
@@ -1,0 +1,177 @@
+<template>
+  <div class="col bar relative-position">
+    <div
+      class="absolute-full scroll"
+      ref="scroll"
+    >
+      <div
+        v-if="fetchingMore"
+        class="full-width text-center generic-padding"
+      >
+        <q-spinner-dots :size="40" />
+      </div>
+      <q-list
+        no-border
+        class="bg-white"
+      >
+        <ConversationMessage
+          v-for="message in reversedMessages"
+          :key="message.id"
+          :message="message"
+          slim
+          @toggleReaction="$emit('toggleReaction', arguments[0])"
+          @save="$emit('saveMessage', arguments[0])"
+        />
+        <ConversationCompose
+          ref="compose"
+          :status="conversation.sendStatus"
+          slim
+          @submit="sendMessage"
+          :placeholder="messagePrompt"
+        />
+      </q-list>
+      <q-scroll-observable @scroll="onScroll" />
+    </div>
+    <q-resize-observable @resize="restoreScrollPosition" />
+  </div>
+</template>
+
+<script>
+import ConversationMessage from '@/components/Conversation/ConversationMessage'
+import ConversationCompose from '@/components/Conversation/ConversationCompose'
+import {
+  scroll,
+  QSpinnerDots,
+  QList,
+  QResizeObservable,
+  QScrollObservable,
+} from 'quasar'
+const { getScrollHeight, getScrollPosition, setScrollPosition } = scroll
+
+export default {
+  components: {
+    ConversationMessage,
+    ConversationCompose,
+    QSpinnerDots,
+    QList,
+    QResizeObservable,
+    QScrollObservable,
+  },
+  props: {
+    conversation: { type: Object, default: null },
+    away: { type: Boolean, required: true },
+    currentUser: { type: Object, default: null },
+  },
+  data () {
+    return {
+      newestMessageId: -1,
+      oldestMessageId: -1,
+      scrollPositionFromBottom: 0,
+    }
+  },
+  computed: {
+    hasLoaded () {
+      if (!this.conversation) return false
+      const s = this.conversation.fetchStatus
+      return !s.pending && !s.hasValidationErrors
+    },
+    messagePrompt () {
+      if (this.conversation.messages.length > 0) {
+        return this.$t('WALL.WRITE_MESSAGE')
+      }
+      else {
+        return this.$t('WALL.WRITE_FIRST_MESSAGE')
+      }
+    },
+    reversedMessages () {
+      if (!this.conversation) return []
+      return this.conversation.messages.slice().reverse()
+    },
+    fetchingMore () {
+      return this.conversation.fetchMoreStatus.pending
+    },
+  },
+  watch: {
+    away (away) {
+      if (!away) this.markRead(this.newestMessageId)
+    },
+    hasLoaded (hasLoaded) {
+      if (hasLoaded) this.scrollToBottom()
+    },
+    'conversation.id' (id) {
+      Object.assign(this, {
+        newestMessageId: -1,
+        oldestMessageId: -1,
+        scrollPositionFromBottom: 0,
+      })
+    },
+    'conversation.messages' (messages) {
+      if (!messages || messages.length === 0) return
+      // Jump to bottom when new messages added
+      const newNewestMessageId = messages[0].id
+      if (this.newestMessageId !== newNewestMessageId) {
+        this.scrollToBottom()
+        this.newestMessageId = newNewestMessageId
+        if (!this.away) this.markRead(this.newestMessageId)
+      }
+      // Retain position when old message added
+      const newOldestMessageId = messages[messages.length - 1].id
+      if (this.oldestMessageId !== newOldestMessageId) {
+        this.saveScrollPosition()
+        this.oldestMessageId = newOldestMessageId
+        this.$nextTick(() => {
+          this.restoreScrollPosition()
+        })
+      }
+    },
+  },
+  methods: {
+    markRead (messageId) {
+      if (!this.conversation) return
+      if (this.conversation.unreadMessageCount > 0) {
+        this.$emit('mark', {
+          id: this.conversation.id,
+          seenUpTo: messageId,
+        })
+      }
+    },
+    sendMessage (content) {
+      this.$emit('send', {
+        id: this.conversation.id,
+        content,
+      })
+    },
+    saveScrollPosition () {
+      if (!this.$refs.scroll) return
+      const { height } = this.$refs.scroll.getBoundingClientRect()
+      const scrollHeight = getScrollHeight(this.$refs.scroll)
+      const scrollPosition = getScrollPosition(this.$refs.scroll)
+      const newScrollPositionFromBottom = scrollHeight - scrollPosition - height
+      if (this.scrollPositionFromBottom !== newScrollPositionFromBottom) {
+        this.scrollPositionFromBottom = newScrollPositionFromBottom
+      }
+    },
+    restoreScrollPosition () {
+      if (!this.$refs.scroll) return
+      const { height } = this.$refs.scroll.getBoundingClientRect()
+      const scrollHeight = getScrollHeight(this.$refs.scroll)
+      const scrollPosition = scrollHeight - height - this.scrollPositionFromBottom
+      setScrollPosition(this.$refs.scroll, scrollPosition)
+    },
+    scrollToBottom () {
+      this.$nextTick(() => {
+        if (this.$refs.scroll) {
+          setScrollPosition(this.$refs.scroll, getScrollHeight(this.$refs.scroll))
+        }
+      })
+    },
+    onScroll ({ position }) {
+      if (position < 50) {
+        if (!this.fetchingMore && this.conversation.canLoadMore) {
+          this.$emit('fetchMore', this.conversation.id)
+        }
+      }
+    },
+  },
+}
+</script>
