+++ src/components/General/DetailUI.vue
@@ -1,15 +1,12 @@
 <template>
   <div class="Detail absolute-full column">
-    <q-alert v-if="conversation && conversation.fetchStatus.hasValidationErrors">
-      {{ conversation.fetchStatus.validationErrors }}
-    </q-alert>
     <div
       v-if="!hasLoaded"
       class="full-width text-center generic-padding"
     >
       <q-spinner-dots :size="40" />
     </div>
-    <template v-if="hasLoaded">
+    <template v-else>
       <div class="col-auto">
         <q-toolbar
           color="secondary"
@@ -70,88 +67,40 @@
           </div>
         </div>
       </div>
-      <div class="col bar relative-position">
-        <div
-          class="absolute-full scroll"
-          ref="scroll"
-        >
-          <div
-            v-if="fetchingMore"
-            class="full-width text-center generic-padding"
-          >
-            <q-spinner-dots :size="40" />
-          </div>
-          <q-list
-            no-border
-            class="bg-white"
-          >
-            <ConversationMessage
-              v-for="message in reversedMessages"
-              :key="message.id"
-              :message="message"
-              slim
-              @toggleReaction="$emit('toggleReaction', arguments[0])"
-              @save="$emit('saveMessage', arguments[0])"
-            />
-            <ConversationCompose
-              ref="compose"
-              :status="conversation.sendStatus"
-              slim
-              @submit="sendMessage"
-              :placeholder="messagePrompt"
-              :user="user"
-            />
-          </q-list>
-          <q-scroll-observable @scroll="onScroll" />
-        </div>
-      </div>
+      <ChatConversation
+        :conversation="conversation"
+        :away="away"
+        :current-user="currentUser"
+        @send="$emit('send', arguments[0])"
+        @mark="$emit('mark', arguments[0])"
+        @toggleReaction="$emit('toggleReaction', arguments[0])"
+        @saveMessage="$emit('saveMessage', arguments[0])"
+        @fetchMore="$emit('fetchMore', arguments[0])"
+      />
     </template>
-    <q-resize-observable @resize="restoreScrollPosition" />
   </div>
 </template>
 
 <script>
 import ProfilePicture from '@/components/ProfilePictures/ProfilePicture'
-import ConversationMessage from '@/components/Conversation/ConversationMessage'
-import ConversationCompose from '@/components/Conversation/ConversationCompose'
 import NotificationToggle from '@/components/Conversation/NotificationToggle'
+import ChatConversation from '@/components/Conversation/ChatConversation'
 import {
-  scroll,
   QBtn,
-  QInfiniteScroll,
-  QSpinnerDots,
-  QList,
-  QAlert,
-  QItem,
-  QIcon,
-  QTooltip,
-  QScrollArea,
   QToolbar,
   QToolbarTitle,
-  QResizeObservable,
-  QScrollObservable,
+  QSpinnerDots,
 } from 'quasar'
-const { getScrollHeight, getScrollPosition, setScrollPosition } = scroll
 
 export default {
   components: {
-    ConversationMessage,
-    ConversationCompose,
+    ChatConversation,
     ProfilePicture,
     NotificationToggle,
     QBtn,
-    QInfiniteScroll,
-    QSpinnerDots,
-    QList,
-    QAlert,
-    QItem,
-    QIcon,
-    QTooltip,
-    QScrollArea,
     QToolbar,
     QToolbarTitle,
-    QResizeObservable,
-    QScrollObservable,
+    QSpinnerDots,
   },
   props: {
     user: { type: Object, default: null },
@@ -160,124 +109,22 @@ export default {
     away: { type: Boolean, required: true },
     currentUser: { type: Object, default: null },
   },
-  data () {
-    return {
-      newestMessageId: -1,
-      oldestMessageId: -1,
-      scrollPositionFromBottom: 0,
-    }
-  },
   computed: {
     hasLoaded () {
       if ((!this.pickup && !this.user) || !this.conversation) return false
       const s = this.conversation.fetchStatus
       return !s.pending && !s.hasValidationErrors
     },
-    messagePrompt () {
-      if (this.conversation.messages.length > 0) {
-        return this.$t('WALL.WRITE_MESSAGE')
-      }
-      else {
-        return this.$t('WALL.WRITE_FIRST_MESSAGE')
-      }
-    },
-    reversedMessages () {
-      if (!this.conversation) return []
-      return this.conversation.messages.slice().reverse()
-    },
-    fetchingMore () {
-      return this.conversation.fetchMoreStatus.pending
-    },
-  },
-  watch: {
-    away (away) {
-      if (!away) this.markRead(this.newestMessageId)
-    },
-    hasLoaded (hasLoaded) {
-      if (hasLoaded) this.scrollToBottom()
-    },
-    'conversation.id' (id) {
-      Object.assign(this, {
-        newestMessageId: -1,
-        oldestMessageId: -1,
-        scrollPositionFromBottom: 0,
-      })
-    },
-    'conversation.messages' (messages) {
-      if (!messages || messages.length === 0) return
-      // Jump to bottom when new messages added
-      const newNewestMessageId = messages[0].id
-      if (this.newestMessageId !== newNewestMessageId) {
-        this.scrollToBottom()
-        this.newestMessageId = newNewestMessageId
-        if (!this.away) this.markRead(this.newestMessageId)
-      }
-      // Retain position when old message added
-      const newOldestMessageId = messages[messages.length - 1].id
-      if (this.oldestMessageId !== newOldestMessageId) {
-        this.saveScrollPosition()
-        this.oldestMessageId = newOldestMessageId
-        this.$nextTick(() => {
-          this.restoreScrollPosition()
-        })
-      }
-    },
   },
   methods: {
     conversationPartner (conversation) {
       return this.conversation && this.conversation.participants && this.conversation.participants.find(e => !e.isCurrentUser)
-    },
-    markRead (messageId) {
-      if (!this.conversation) return
-      if (this.conversation.unreadMessageCount > 0) {
-        this.$emit('mark', {
-          id: this.conversation.id,
-          seenUpTo: messageId,
-        })
-      }
-    },
-    sendMessage (content) {
-      this.$emit('send', {
-        id: this.conversation.id,
-        content,
-      })
     },
     toggleNotifications () {
       this.$emit('toggleEmailNotifications', {
         conversationId: this.conversation.id,
         value: !this.conversation.emailNotifications,
       })
-    },
-    saveScrollPosition () {
-      if (!this.$refs.scroll) return
-      const { height } = this.$refs.scroll.getBoundingClientRect()
-      const scrollHeight = getScrollHeight(this.$refs.scroll)
-      const scrollPosition = getScrollPosition(this.$refs.scroll)
-      const newScrollPositionFromBottom = scrollHeight - scrollPosition - height
-      if (this.scrollPositionFromBottom !== newScrollPositionFromBottom) {
-        this.scrollPositionFromBottom = newScrollPositionFromBottom
-      }
-    },
-    restoreScrollPosition () {
-      if (!this.$refs.scroll) return
-      const { height } = this.$refs.scroll.getBoundingClientRect()
-      const scrollHeight = getScrollHeight(this.$refs.scroll)
-      const scrollPosition = scrollHeight - height - this.scrollPositionFromBottom
-      setScrollPosition(this.$refs.scroll, scrollPosition)
-    },
-    scrollToBottom () {
-      this.$nextTick(() => {
-        if (this.$refs.scroll) {
-          setScrollPosition(this.$refs.scroll, getScrollHeight(this.$refs.scroll))
-        }
-      })
-    },
-    onScroll ({ position }) {
-      if (position < 50) {
-        if (!this.fetchingMore && this.conversation.canLoadMore) {
-          this.$emit('fetchMore', this.conversation.id)
-        }
-      }
     },
   },
 }
@@ -287,10 +134,6 @@ export default {
 @import '~variables'
 .Detail
   background-color white
-.actionButton
-  float right
-  margin-top -25px
-  margin-right 5px
 .k-participant-list
   background-color #f5f5f5
   padding 0.3em
